Corrección en generación y validación de token

+++ BibliotecaApi/Models/Autor.cs
@@ -4,8 +4,8 @@ namespace BibliotecaApi.Models
 {
     public class Autor
     {
-        public long Id { get; set; }
-        public long PersonaId { get; set; }
+        public int Id { get; set; }
+        public int PersonaId { get; set; }
 
         public Persona? Persona { get; set; }
     }
