+++ BibliotecaApi/Models/AutorLibro.cs
@@ -4,9 +4,9 @@ namespace BibliotecaApi.Models
 {
     public class AutorLibro
     {
-        public long Id { get; set; }
-        public long AutorId { get; set; }
-        public long LibroId { get; set; }
+        public int Id { get; set; }
+        public int AutorId { get; set; }
+        public int LibroId { get; set; }
 
         public Libro? Libro { get; set; }
         public Autor? Autor { get; set; }
